+++ ricsdl-package/ricsdl/__init__.py
@@ -31,7 +31,7 @@ from .exceptions import (
 )
 
 
-__version__ = '2.3.0'
+__version__ = '3.0.0'
 
 
 __all__ = [